Add fs dependency tracker
The problem is that filesystems can depend on volumes which depend
on other volumes, which can in turn be based on a filesystem and
once we call build_format() on a file system, the result is final
and immutable. We need to make sure the build_format() and commit()
calls are made in the correct dependency order.

This also crates a new library called 'libimgtool.a' which contains
the higher level application logic.

Signed-off-by: David Oberhollenzer <goliath@infraroot.at>