Clarify the licensing situation
Add a copy of the GPLv3 and replace switch the rest of the gcfg
code over to the GPL to simplify things.

Since I'm the sole copyright holder of the gcfg library, I can do
this. For the combined work, the GPL applies anyway and if someone
ripps one of the unmodified files out, they can still claim that
they obtained it elsewhere under the terms and conditions of the
ISC license.

The only downside of this is that third party contributions that
change code of the libgcfg library can no longer flow back into
the original gcfg code.

Signed-off-by: David Oberhollenzer <goliath@infraroot.at>