Fix: fatfs empty root directory special case handling
We normally don't need to worry about empty directories, because
they are never really empty, but instead contain a "." and ".."
entry, but the root directory doesn't and really is 0 bytes in
size.

If this case isn't handled specifically, we end up having an invalid
cluster chain for the root directory.

Signed-off-by: David Oberhollenzer <goliath@infraroot.at>