Fix: fstree: only initialize link target if extra is not NULL
The code previously assumed we have extra != NULL for symlinks and
hard links, and first checked if extra was NULL when allocating
space for it. Theoretically we could run into a case where extra
is erroneously NULL for a link, where we would allocate nothing and
then try to copy from the NULL pointer to somehwere after the
allocated buffer.

This commit fixes the mknode_at function by returning an error if
when trying to create a symlink or hard link without a target.

Signed-off-by: David Oberhollenzer <goliath@infraroot.at>