Fix listing file source decoding of longer entries
Entries >= 32 characters are stored by setting the 5 bit length
field to 0 and storing the actual length - 32 in a full byte before
the name string.

The check as copy & pasted from the encoding is of course wrong in
checking whether namelen > 31 and should instead test if it is 0
before fetching the actual length.

Signed-off-by: David Oberhollenzer <goliath@infraroot.at>