Add simple implementation of MBR volume partitioning
A new "partition manager" type is added to libimage.a that can
create partition sub-volumes from a parent volume. The volume is
not exposed externally to allow it to possibly work on more than
one volume.

A simple MBR implementation is added that prefixes up to four
primary partitions with a DOS MBR.

Signed-off-by: David Oberhollenzer <goliath@infraroot.at>